fix: resolve MetaTask field mismatch and ensure seed overrides persist
This commit fixes the 'cargo test' failures caused by missing 'seed_ratio_override' and 'seed_time_override' fields in MetaTask struct initializations across multiple test files. It also ensures that these overrides are correctly persisted in the MetaTask and applied when registering BitTorrent tasks. Additionally, refactored dependency logic to keep files within maintainability limits.
