+++ aura-core/src/orchestrator/tests_dependencies.rs
@@ -32,6 +32,8 @@ async fn test_dependency_cycle_detection() {
         etag: None,
         last_modified: None,
         selected_files: None,
+        seed_ratio_override: None,
+        seed_time_override: None,
     };
     orch.tasks.insert(TaskId(1), meta_a);
 
@@ -59,6 +61,8 @@ async fn test_dependency_cycle_detection() {
         etag: None,
         last_modified: None,
         selected_files: None,
+        seed_ratio_override: None,
+        seed_time_override: None,
     };
     orch.tasks.insert(TaskId(2), meta_b);
 
@@ -86,6 +90,8 @@ async fn test_dependency_cycle_detection() {
         etag: None,
         last_modified: None,
         selected_files: None,
+        seed_ratio_override: None,
+        seed_time_override: None,
     };
     orch.tasks.insert(TaskId(3), meta_c);
 
@@ -127,6 +133,8 @@ async fn test_dependency_waiting_state_and_unblocking() {
         etag: None,
         last_modified: None,
         selected_files: None,
+        seed_ratio_override: None,
+        seed_time_override: None,
     };
     orch.tasks.insert(TaskId(1), meta_a);
 
@@ -189,6 +197,8 @@ async fn test_follow_on_custom_trigger() {
         etag: None,
         last_modified: None,
         selected_files: None,
+        seed_ratio_override: None,
+        seed_time_override: None,
     };
     orch.tasks.insert(meta_id, meta);
